+++ src/main/java/com/github/jlangch/venice/impl/repl/TerminalPrinter.java
@@ -21,13 +21,14 @@
  */
 package com.github.jlangch.venice.impl.repl;
 
+import static com.github.jlangch.venice.impl.util.StringUtil.truncate;
+
 import java.util.function.Consumer;
 
 import org.jline.terminal.Terminal;
 
 import com.github.jlangch.venice.ValueException;
 import com.github.jlangch.venice.VncException;
-import com.github.jlangch.venice.impl.Printer;
 import com.github.jlangch.venice.impl.types.VncVal;
 
 
@@ -97,14 +98,21 @@ public class TerminalPrinter {
         synchronized (lock) {
             try {
                 if (ex instanceof ValueException) {
-                    print(
-                        colorID,
-                        t -> ((ValueException)ex).printVeniceStackTrace(t.writer()));
-                    println(
-                        colorID,
-                        "\nThrown value: " + Printer.pr_str(
-                                                (VncVal)((ValueException)ex).getValue(),
-                                                false));
+                    final ValueException vex = (ValueException)ex;
+                    final Object value = vex.getValue();
+                    final String fmtVal;
+                    if (value == null) {
+                        fmtVal = "null";
+                    }
+                    else if (value instanceof VncVal) {
+                        fmtVal = ((VncVal)value).toString(true);
+                    }
+                    else {
+                        fmtVal = value.toString();
+                    }
+
+                    print(colorID, t -> vex.printVeniceStackTrace(t.writer()));
+                    println(colorID,"\nThrown value: " + truncate(fmtVal, 200, "...(skipped)"));
                 }
                 else if (ex instanceof VncException) {
                     if (printJavaEx) {